Modify assigment7_1_2
-Adding improvements
